init: add skeleton of project
This project aims to provide a means to select an image size
and number of layers, and programmatically build across a matrix
of size by number of layers. This is not a huge project, but I see
it necessary to properly run experiments that test this relationship
for containers in Kubernetes in the cloud.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>